+++ NEWS.md
@@ -3,10 +3,13 @@
 ## Enhancements
 
 - Re-factored header layout to be more flexible with menu items
+- Updated modules default values
+- **downloadableTable** module server parameters "filenameroot" can now support character,  reactive expression or a function to return file name
 
 ## Bug Fixes
 
-- Hide download button if `downloaddatafxn` or `downloadtypes` are empty and updated module default values
+- Hide download button if `downloaddatafxn` or `downloadtypes` are empty for `downloadableTable` module
+- Hide download button if `downloadfxns` or `downloadtypes` are empty for `downloadablePlot` module
 
 -----